+++ crates/amaru-ouroboros-traits/src/has_stake_pools.rs
@@ -15,17 +15,21 @@
 use std::collections::BTreeSet;
 
 use amaru_kernel::PeerCandidate;
+use async_trait::async_trait;
 
 use crate::BlockValidationError;
 
+#[expect(clippy::double_must_use)]
+#[async_trait]
 pub trait HasStakePools: Send + Sync {
-    fn registered_relay_candidates(&self) -> Result<BTreeSet<PeerCandidate>, BlockValidationError>;
+    async fn registered_relay_candidates(&self) -> Result<BTreeSet<PeerCandidate>, BlockValidationError>;
 }
 
 pub struct MockHasStakePools;
 
+#[async_trait]
 impl HasStakePools for MockHasStakePools {
-    fn registered_relay_candidates(&self) -> Result<BTreeSet<PeerCandidate>, BlockValidationError> {
+    async fn registered_relay_candidates(&self) -> Result<BTreeSet<PeerCandidate>, BlockValidationError> {
         Ok(Default::default())
     }
 }